chore: move elements to elements dir
