+++ elements/MarqueeText/FadeShader.js
