+++ elements/MarqueeText/index.js
